Issue #342: Don't call docker stop if no running containers
There was errors in the log:
"docker stop" requires at least 1 argument.
"docker rm" requires at least 1 argument.

+++ jenkins_config/build.pl
@@ -108,8 +108,14 @@ sub run {
 
 sub docker_cleanup {
     run(q{docker-compose -p koha down});
-    run(qq{docker stop \$(docker ps -a -f "name=koha_" -q)});
-    run(qq{docker rm \$(docker ps -a -f "name=koha_" -q)});
+
+    my $containers = qx{docker ps -a -f "name=koha_xx" -q};
+    chomp $containers;
+    if ( $containers ) {
+        run(qq{docker stop \$(docker ps -a -f "name=koha_" -q)});
+        run(qq{docker rm \$(docker ps -a -f "name=koha_" -q)});
+    }
+
     run(q{docker volume prune -f});
     run(q{docker image  prune -f});
     run(q{docker system prune -a -f});
